+++ wulu/src/index.js
@@ -3,6 +3,8 @@ import Commands from './build/commands';
 import Economy from './build/economy';
 import {Emoticons} from './build/emoticons';
 import Mongo from './build/mongo';
+import Poll from './build/poll';
+import Sysop from './build/sysop';
 import Tour from './build/tour';
 
 export default {
@@ -11,5 +13,7 @@ export default {
   Economy,
   Emoticons,
   Mongo,
+  Poll,
+  Sysop,
   Tour
 };